Multiprocessing in rmgpy.pdep.network
For each Network.calculateRateCoefficients() call, we create a new 
pool of workers, initialized with a copy of the network. The array
of temperatures and pressures at which the network needs to be solved
is then divided up amongst the workers in the pool. The rate coefficients
are returned, stored in the K array as before.

I'm not sure what overhead is for creating and destroying the pool of workers
so often. There's probably a smarter way to keep them around for longer
and just send them the new network info as needed, but this seems to function.

NB. this commit is independent of others on the multiprocessing branch 
and could (probably) be implemented on its own.