Deal with explicit reverse reactions differently when reading chemkin files.
Instead of just skipping any line containing REV we now attempt to read it.
We store the explicit rate in a comment on the kinetics, and if it
is zero, we make the reaction irreversible. We do not create explicit reverse
reactions (just store the comment).
I expect it will not work for PDep or Duplicate reactions, only plain vanilla
high-P Arrhenius rates.